modify run way

+++ PageRank/src/main/scala/org/ymt/spark/graphx/pageRank/Rank.scala
@@ -15,7 +15,7 @@ object Rank extends Serializable {
     val inputPath = args(0)
     val outputPath = args(1)
     val numPartitions = args(2).toInt
-    val rank = args(3)
+    val rank = args(3).toDouble
     // graph loader phase
     val graph = GraphLoader.edgeListFile(sc, inputPath).cache()
     val g = Graph(graph.vertices.repartition(numPartitions),
